analyzer: Move "analyze" function from Main to its own class
...to separate analyzer logic from the CLI code.
