+++ analyzer/src/main/kotlin/Analyzer.kt
@@ -1,0 +1,99 @@
+/*
+ * Copyright (C) 2017-2018 HERE Europe B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package com.here.ort.analyzer
+
+import ch.frankel.slf4k.*
+
+import com.here.ort.analyzer.managers.Unmanaged
+import com.here.ort.downloader.VersionControlSystem
+import com.here.ort.model.AnalyzerConfiguration
+import com.here.ort.model.AnalyzerResultBuilder
+import com.here.ort.model.ProjectAnalyzerResult
+import com.here.ort.utils.log
+
+import java.io.File
+
+class Analyzer {
+    fun analyze(config: AnalyzerConfiguration, absoluteProjectPath: File,
+                packageManagers: List<PackageManagerFactory<PackageManager>> = PackageManager.ALL,
+                packageCurationsFile: File? = null
+    ): AnalyzerResultBuilder {
+        // Map of files managed by the respective package manager.
+        val managedDefinitionFiles = if (packageManagers.size == 1 && absoluteProjectPath.isFile) {
+            // If only one package manager is activated, treat the given path as definition file for that package
+            // manager despite its name.
+            mutableMapOf(packageManagers.first() to listOf(absoluteProjectPath))
+        } else {
+            PackageManager.findManagedFiles(absoluteProjectPath, packageManagers).toMutableMap()
+        }
+
+        val hasDefinitionFileInRootDirectory = managedDefinitionFiles.values.flatten().any {
+            it.parentFile.absoluteFile == absoluteProjectPath
+        }
+
+        if (managedDefinitionFiles.isEmpty() || !hasDefinitionFileInRootDirectory) {
+            managedDefinitionFiles[Unmanaged] = listOf(absoluteProjectPath)
+        }
+
+        if (log.isInfoEnabled) {
+            // Log the summary of projects found per package manager.
+            managedDefinitionFiles.forEach { manager, files ->
+                // No need to use curly-braces-syntax for logging here as the log level check is already done above.
+                log.info("$manager projects found in:")
+                log.info(files.joinToString("\n") {
+                    "\t${it.toRelativeString(absoluteProjectPath).let { if (it.isEmpty()) "." else it }}"
+                })
+            }
+        }
+
+        val vcs = VersionControlSystem.getCloneInfo(absoluteProjectPath)
+        val analyzerResultBuilder = AnalyzerResultBuilder(config, vcs)
+
+        // Resolve dependencies per package manager.
+        managedDefinitionFiles.forEach { manager, files ->
+            val results = manager.create(config).resolveDependencies(absoluteProjectPath, files)
+
+            val curatedResults = packageCurationsFile?.let {
+                val provider = YamlFilePackageCurationProvider(it)
+                results.mapValues { entry ->
+                    ProjectAnalyzerResult(
+                            project = entry.value.project,
+                            errors = entry.value.errors,
+                            packages = entry.value.packages.map { curatedPackage ->
+                                val curations = provider.getCurationsFor(curatedPackage.pkg.id)
+                                curations.fold(curatedPackage) { cur, packageCuration ->
+                                    log.debug {
+                                        "Applying curation '$packageCuration' to package '${curatedPackage.pkg.id}'."
+                                    }
+                                    packageCuration.apply(cur)
+                                }
+                            }.toSortedSet()
+                    )
+                }
+            } ?: results
+
+            curatedResults.forEach { _, analyzerResult ->
+                analyzerResultBuilder.addResult(analyzerResult)
+            }
+        }
+
+        return analyzerResultBuilder
+    }
+}
